chore(assumptions): simplify return type of `incremental_push_all`
The ordering of the `Vec<bool>` result could be problematic if the input `impl IntoIterator<Item = Lit>` is a collection that, if cloned, wouldn't necessarily be iterated in the same order (like a `HashMap`). This could cause confusion or bugs. If one wants the boolean result of making each assumption (true if it was new, false if already held), one should use `incremental_push`, not `incremental_push_all`.