fix(marco): refactor marco w/ improvements
Most notably, a `solve_fn` closure is given as parameter to `run` method of Marco, and is mostly used in the subset solver. This will probably be useful when we try to use Marco with planning problems, where the solving is actually a bit more complicated than simply using the basic constraint problem as is.

For the simple case where only a "standard" constraint satisfactino problem is addressed, the `SimpleMarco` is to be used (it is a thin wrapper over Marco)