fix(osp): include `Solver` rather than `Model` in `EncodedProblem`
This caused a bug with assumptions.
The function `encode` posted constraints and propagated in a temporary solver, and then only outputted its model. When a new solver object was created later with that same model, the posted constraints were forgotten, which resulted in new constraints being posted. Somehow, this sometimes caused a bug when explaining conflicts during the solving procedure, with some literals' cause / entailing event pointing to an incorrect (non-existent / outdated clause).